package/gnuradio: bump to 3.9.3.0
This adds preliminary support for GNU Radio 3.9

There are a number of things that have been done in a far from
ideal way in order to get GNU Radio to build in a haste.

The detection of numpy has been disabled, since it was failing even
though numpy was built and installed to the STAGING_DIR. This also
requires the include path of the numpy headers to be set by hand
in gnuradio.mk.

A weird problem with a static constexpr would create an undefined
symbol in the libgnuradio-digital.so has been fixed by replacing
the static constexpr with a #define

Moreover, after installation, the C++ Python bindings dynamic
libraries are missing the .so extension, so they need to be renamed
(this fix isn't included in this commit).

This commit also adds a package for volk, since it is now distributed
separately from GNU Radio.